requirements.txt
websocket
PyEventEmitter